Get all route and db
Fixes #28
Fixes #27

Some things:
- Mongo wants when searching on the fields of a sub-document for the structure
to be field.value. I'm taking the params and setting them to a map and then
taking that map and restructuring it for mongo. Its possible the first map
will not be needed but I'll keep it for now because it might be required
somewhere else
- Not including any params correctly gets all documents